Force lower-case node names
Fixes a bug when submitting a node with a capitalized node name
(e.g. HOSTNAME.domain.local). The Foreman downcases node names when
creating a new node from a fact import, but does not follow the same
logic when creating a node from a report import. This leads to two
nodes being created with one containing the reports, and
the other containing the facts. See the following bug issue:
http://projects.theforeman.org/issues/12185